Read users.gh_id as int64 instead of int
int64 is the type of user.ID in the go-github library, so we might as
well read gh_id from the database as an int64 to avoid casting every
time we want to compare it to a user.ID.

+++ api/publish_app.go
@@ -11,7 +11,7 @@ import (
 
 func PublishApp(c *gin.Context) {
 	db := c.MustGet("db").(*sql.DB)
-	ghID := c.MustGet("gh_id").(int)
+	ghID := c.MustGet("gh_id").(int64)
 	appID := c.Param("appID")
 
 	tx, err := db.Begin()
